remove impl from service logic
