+++ project_seperate/src/service/turn_execute_service.rs
@@ -1,41 +1,15 @@
 use rusqlite::Connection;
 
 use crate::repository::{
-    player_repo::{update_money, update_position_and_lap, bankrupt},
-    property_repo::{set_owner, reset_owner_for_player},
+    player_repo::{update_money, update_position_and_lap},
+    property_repo::{set_owner},
     transcaction_repo::record_transaction,
 };
 
 use crate::service::turn_service::{TurnResult, TurnAction};
 use crate::service::traits::TurnExecuteRepo;
+use crate::service::port_impl::PortImpl;
 
-pub struct TurnExecuteRepoImpl;
-
-impl TurnExecuteRepo for TurnExecuteRepoImpl {
-    fn update_position_and_lap(&self, conn: &Connection, player_id: i32, pos: i32, lap: i32) -> rusqlite::Result<()> {
-        update_position_and_lap(conn, player_id, pos, lap)
-    }
-    fn update_money(&self, conn: &Connection, player_id: i32, delta: i32) -> rusqlite::Result<()> {
-        update_money(conn, player_id, delta)
-    }
-    fn record_transaction(&self, conn: &Connection, player_id: i32, tx_type: &str, amount: i32, target: &str) -> rusqlite::Result<()> {
-        record_transaction(conn, player_id, tx_type, amount, target)
-    }
-    fn reset_owner_for_player(&self, conn: &Connection, player_id: i32) -> rusqlite::Result<()> {
-        reset_owner_for_player(conn, player_id)
-    }
-    fn bankrupt(&self, conn: &Connection, player_id: i32) -> rusqlite::Result<()> {
-        bankrupt(conn, player_id)
-    }
-    fn add_fund(&self, conn: &Connection, amount: i32) -> rusqlite::Result<()> {
-        use crate::repository::event_repo::add_fund;
-        add_fund(conn, amount)
-    }
-    fn reset_fund(&self, conn: &Connection) -> rusqlite::Result<()> {
-        use crate::repository::event_repo::reset_fund;
-        reset_fund(conn)
-    }
-}
 
 // process_turn 함수 실행 결과를 DB에 반영하는 함수 (DI 버전)
 pub fn apply_turn_result_with_repo<R: TurnExecuteRepo>(
@@ -212,7 +186,7 @@ pub fn apply_turn_result(
     player_id: i32,
     result: &TurnResult,
 ) -> rusqlite::Result<()> {
-    apply_turn_result_with_repo(&TurnExecuteRepoImpl, conn, player_id, result)
+    apply_turn_result_with_repo(&PortImpl, conn, player_id, result)
 }
 
 /// 이동 + 월급만 선반영 (구매 결정 대기 시 사용)
